CA: Copy the requested "CN" into the SAN (closes #3)
I'm not sure if this only happens on the client side. But currently it
seems that the server is happy with the certs we use for the clients.

All our certs issued so far do not have the SAN set. (Except for the
cert-proxy-server instances, there I created a new cert.)